+++ src/bl_loader/test/python/test_behavior.py
@@ -8,15 +8,18 @@ class TestBehavior(Behavior):
 
     Multiplies every input by a value specified in the meta parameters.
     """
-    def __init__(self, num_inputs, num_outputs, test_value_a, test_value_b):
-        super(TestBehavior, self).__init__(num_inputs, num_outputs)
+    def __init__(self, test_value_a, test_value_b):
         #test_value_a/b are hard coded in init.yaml
         self.can_step_counter = 5;
         assert test_value_a == 42.0; 
         assert test_value_b == 44.0;
-        assert num_inputs == num_outputs
-        self.inputs = np.empty(3)
-        self.outputs = np.empty(3)
+
+    def init(self, n_inputs, n_outputs):
+        assert n_inputs == n_outputs
+        self.n_inputs = n_inputs
+        self.n_outputs = n_outputs
+        self.inputs = np.empty(n_inputs)
+        self.outputs = np.empty(n_outputs)
 
     def set_meta_parameters(self, keys, values):
         assert len(keys) == len(values)
@@ -25,9 +28,9 @@ class TestBehavior(Behavior):
         self.expected_input = meta_parameters["expected_input"]
         assert "multiplier" in meta_parameters
         self.multiplier = meta_parameters["multiplier"]
-        assert len(self.expected_input) == self.num_inputs
 
     def set_inputs(self, inputs):
+        assert len(self.expected_input) == self.n_inputs
         assert_array_equal(inputs, self.expected_input)
         self.inputs[:] = inputs