+++ NEWS.md
@@ -8,3 +8,6 @@ the University of Bremen, we changed the following details:
 * Python interface: `ContextualEnvironment` is now a subclass of `Environment`
 * Python interface: renamed `Environment.get_maximal_feedback` to
   `Environment.get_maximum_feedback`
+* Python and C++ interface: `Behavior` constructor does not take any arguments,
+  instead the function `Behavior.init(int, int)` has been introduced to
+  determine the number of inputs and outputs and initialize the behavior